refactor texture file
