+++ src/assets/textures/utils.js
@@ -1,28 +1,17 @@
-import { Cache } from 'pixi.js';
 import { TextureStyle } from '../../display/data-schema/component-schema';
-import { getAsset } from '../utils';
-import { createRectTexture } from './rect';
+import { renderer } from '../../renderer';
+import { deepMerge } from '../../utils/deepmerge/deepmerge';
 
-export const getTexture = (config) => {
-  let texture = null;
-  if (typeof config === 'string') {
-    texture = getAsset(config);
-  } else {
-    texture = Cache.has(cacheKey(config))
-      ? getAsset(cacheKey(config))
-      : createTexture(config);
-  }
-  return texture;
-};
+const RESOLUTION = 5;
 
-export const createTexture = (config) => {
-  let texture = null;
-  switch (config.type) {
-    case 'rect':
-      texture = createRectTexture(config);
-      break;
-  }
-  Cache.set(cacheKey(config), texture);
+export const generateTexture = (target = null, opts = {}) => {
+  const options = deepMerge({ resolution: RESOLUTION }, opts);
+  if (!target) return;
+
+  const texture = renderer.generateTexture({
+    target,
+    resolution: options.resolution,
+  });
   return texture;
 };
